requirements.txt
Django
django-bootstrap3
django-jquery
djangorestframework
psycopg2==2.8.4
psycopg2-binary==2.8.4
drf-yasg
django-filter